[REFACTOR] firebase로 연결
