+++ client/src/queries/postDetailQuery.js
@@ -1,11 +1,11 @@
-import { getPost } from '../api/post';
+import { getPost } from '../../firebase/posts';
 
 const staleTime = 3000;
 
 const postDetailQuery = postId => ({
   queryKey: ['postDetail', postId],
   queryFn: async () => {
-    const { data } = await getPost(postId);
+    const data = await getPost({ postId });
     return data;
   },
   staleTime,
